refactor(connect-common): move pathUtils out of @trezor/connect
Suite, connect-popup and trading deep-imported
@trezor/connect/src/utils/pathUtils, which slows down TypeScript and was
tracked for extraction in #27376. pathUtils' whole dependency footprint
(ERRORS, PROTO, CoinInfo, DerivationPath) already lives in
@trezor/connect-common, so that is its natural home.

- Move pathUtils and its unit test into @trezor/connect-common and add the
  @trezor/crypto-utils dependency it needs for the HD-path primitives.
- Export the three consumer-facing helpers (getSerializedPath,
  getSlip44ByPath, validatePath) from the connect-common public barrel; the
  rest stays connect-internal to keep the barrel minimal.
- Keep a thin re-export shim in @trezor/connect so its internal call sites
  are unchanged.
- Repoint the external consumers to the @trezor/connect-common barrel and
  drop the eslint-disable exceptions.

The deprecated __btcUnknownTxDebug__ diagnostic moves along with pathUtils
and keeps its existing (deep) import in wallet-core; it will be removed
separately later.
