+++ packages/connect-common/src/index.ts
@@ -17,4 +17,5 @@ export {
 } from './data/connectSettings';
 export * from './utils/debug';
 export * from './utils/urlUtils';
+export { getSerializedPath, getSlip44ByPath, validatePath } from './utils/pathUtils';
 export { connectCallableMethods } from './callableMethods';
