+++ packages/connect-common/src/utils/__tests__/pathUtils.test.ts
