+++ packages/connect-common/tsconfig.lib.json
@@ -7,6 +7,7 @@
         "types": ["node", "chrome", "jest"]
     },
     "references": [
+        { "path": "../crypto-utils" },
         { "path": "../device-utils" },
         { "path": "../protobuf" },
         { "path": "../protocol" },
