+++ suite-common/connect-popup/src/methodHooks/ethereumSignTransaction.ts
@@ -13,8 +13,7 @@ import type {
     EthereumSignTypedData,
     MethodInfo,
 } from '@trezor/connect';
-// eslint-disable-next-line @typescript-eslint/no-restricted-imports -- TODO: extract pathUtils to a shared location and remove this exception (see #27376 deferred work)
-import { getSerializedPath, validatePath } from '@trezor/connect/src/utils/pathUtils';
+import { getSerializedPath, validatePath } from '@trezor/connect-common';
 
 import { connectPopupActions } from '../connectPopupActions';
 import { createPlaceholderAccount } from './utils';
