+++ suite-common/trading/src/thunks/common/createPaymentRequestsThunk.ts
@@ -9,8 +9,7 @@ import { createThunk } from '@suite-common/redux-utils';
 import { selectAccountByKey } from '@suite-common/wallet-core';
 import { type Account, type GeneralPrecomposedTransaction } from '@suite-common/wallet-types';
 import { type PROTO } from '@trezor/connect';
-// eslint-disable-next-line @typescript-eslint/no-restricted-imports -- TODO: extract pathUtils to a shared location and remove this exception (see #27376 deferred work)
-import { getSlip44ByPath, validatePath } from '@trezor/connect/src/utils/pathUtils';
+import { getSlip44ByPath, validatePath } from '@trezor/connect-common';
 import { exhaustive } from '@trezor/type-utils';
 
 import { getNonce } from './getNonce';
