+++ suite-common/trading/src/utils/signature/signatureUtils.ts
@@ -9,8 +9,7 @@ import { toChecksumAddress } from '@suite-common/address';
 import type { Network } from '@suite-common/wallet-config';
 import { asAmountUnit, unitsToSubunits } from '@suite-common/wallet-utils';
 import { type PROTO } from '@trezor/connect';
-// eslint-disable-next-line @typescript-eslint/no-restricted-imports -- TODO: extract pathUtils to a shared location and remove this exception (see #27376 deferred work)
-import { validatePath } from '@trezor/connect/src/utils/pathUtils';
+import { validatePath } from '@trezor/connect-common';
 import { BigNumber, formatBigUintToLE } from '@trezor/utils';
 
 import { cryptoIdToNetworkAndContractAddress } from '../../utils';
